requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anndata==0.10.9
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.9.1
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
biom-format==2.1.16
biopython==1.84
bleach==6.2.0
bokeh==3.4.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cliffs-delta==1.0.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
deepspeed==0.5.9
defusedxml==0.7.1
dgl==2.1.0
DLLogger @ git+https://github.com/NVIDIA/dllogger.git@0540a43971f4a8a16693a9de9de73c1072020769
dm-tree==0.1.8
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fair-esm==2.0.0
fastjsonschema==2.20.0
filelock==3.11.0
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
get-annotations==0.1.2
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface==0.0.1
idna==3.10
igraph==0.11.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json-tricks==3.17.3
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
legacy-api-wrap==1.4
Levenshtein==0.27.1
lightning-utilities==0.11.8
livelossplot==0.5.6
llvmlite==0.43.0
logomaker==0.8.6
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
ml_collections==0.1.1
mpmath==1.3.0
multidict==6.1.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
nni==3.0
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
polyleven==0.8
prettytable==3.12.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptflops==0.7.4
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py3Dmol==2.4.2
pycparser==2.22
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
PySnooper==1.2.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-Levenshtein==0.27.1
PythonWebHDFS==0.2.3
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
responses==0.25.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scanpy==1.10.3
schema==0.7.7
scikit-bio==0.6.3
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
session_info==1.0.0
shap==0.46.0
simplejson==3.19.3
six==1.16.0
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
stdlib-list==0.11.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.0.2
torch==2.5.1+cu121
torch-geometric==2.6.1
torchaudio==2.5.1+cu121
torchdata==0.9.0
torchmetrics==1.5.1
TorchSnooper==0.8
torchvision==0.20.1+cu121
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
triton==3.1.0
typeguard==4.1.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.2.3
ViennaRNA==2.7.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xyzservices==2025.1.0
yarl==1.17.0
zipp==3.20.2
